fix: renamed whychoose file
